clippy: Remove unnecessary type casts
Address clippy::unnecessary_cast by removing redundant `as u32` casts
on constants and return values, and removing unnecessary `try_into().unwrap()`.

+++ scheds/rust/scx_forge/src/main.rs
@@ -788,7 +788,7 @@ impl<'a> Scheduler<'a> {
         };
         let out = prog.test_run(input).unwrap();
         if out.return_value != 0 {
-            return Err(out.return_value as u32);
+            return Err(out.return_value);
         }
         Ok(())
     }
